requirements.txt
Flask==3.1.0
PyMuPDF==1.25.3
pdfplumber==0.11.6
google-genai==1.14.0
python-dotenv==1.1.0
Werkzeug==3.1.3
